add mob summons

+++ channel/pools.go
@@ -250,7 +250,7 @@ func (pool *lifePool) mobAcknowledge(poolID int32, plr *Player, moveID int16, sk
 				debuffSkillID, debuffSkillLevel, debuffSkillData := pool.mobs[i].performSkill(skillDelay, skillLevel, skillID)
 				// Apply debuffs to players in range if skill returned debuff info
 				if debuffSkillID != 0 {
-					pool.applyMobDebuffToPlayers(mob, debuffSkillID, debuffSkillLevel, debuffSkillData)
+					pool.performSkill(mob, debuffSkillID, debuffSkillLevel, debuffSkillData)
 				}
 			} else if actualAction > 12 && actualAction < 20 {
 				attackID := byte(actualAction - 12)
@@ -290,15 +290,13 @@ func (pool *lifePool) mobAcknowledge(poolID int32, plr *Player, moveID int16, sk
 	}
 }
 
-func (pool *lifePool) applyMobDebuffToPlayers(mob *monster, skillID, skillLevel byte, skillData nx.MobSkill) {
+func (pool *lifePool) performSkill(mob *monster, skillID, skillLevel byte, skillData nx.MobSkill) {
 	if pool.instance == nil {
 		return
 	}
 
-	// Handle special skills first
 	switch skillID {
 	case skill.Mob.Dispel:
-		// Dispel removes all buffs from players
 		for _, plr := range pool.instance.players {
 			if plr == nil || plr.buffs == nil {
 				continue
@@ -307,7 +305,6 @@ func (pool *lifePool) applyMobDebuffToPlayers(mob *monster, skillID, skillLevel 
 		}
 		return
 	case skill.Mob.HealAoe:
-		// Heal all mobs in the area
 		healAmount := skillData.Hp
 		for _, m := range pool.mobs {
 			if m != nil {
@@ -334,22 +331,21 @@ func (pool *lifePool) applyMobDebuffToPlayers(mob *monster, skillID, skillLevel 
 	case skill.Mob.MagicImmunity:
 		mob.statBuff |= skill.MobStat.MagicImmune
 		return
+	case skill.Mob.Summon:
+		pool.handleMobSummon(mob, skillLevel, skillData)
+		return
 	}
 
-	// Get all players in the field instance
 	for _, plr := range pool.instance.players {
 		if plr == nil || plr.buffs == nil {
 			continue
 		}
 
-		// Apply the debuff
-		// The duration is in the skill data Time field (in seconds)
 		durationSec := int16(0)
 		if skillData.Time > 0 {
 			durationSec = int16(skillData.Time) // Time is already in seconds
 		}
 
-		// Use AddMobDebuff which we'll add to CharacterBuffs
 		plr.addMobDebuff(skillID, skillLevel, durationSec)
 	}
 }
@@ -370,39 +366,39 @@ func (pool *lifePool) mobDamaged(poolID int32, damager *Player, dmg ...int32) {
 			pool.showMobBossHPBar(v, nil)
 
 			if pool.mobs[i].hp < 1 {
-				for plr, dmg := range pool.mobs[i].dmgTaken {
-					if damager != nil && damager.mapID != plr.mapID {
-						continue
-					}
-
-					var partyExp int32
-
-					if dmg == v.maxHP {
-						plr.giveEXP(v.exp, true, false)
-						partyExp = int32(float64(v.exp) * 0.25) // TODO: party exp needs to be properly calculated
-					} else if float64(dmg)/float64(v.maxHP) > 0.60 {
-						plr.giveEXP(v.exp, true, false)
-						partyExp = int32(float64(v.exp) * 0.25) // TODO: party exp needs to be properly calculated
-					} else {
-						newExp := int32(float64(v.exp) * 0.25)
-
-						if newExp == 0 {
-							newExp = 1
+				if damager != nil {
+					for plr, dmg := range pool.mobs[i].dmgTaken {
+						if damager.mapID != plr.mapID {
+							continue
 						}
 
-						plr.giveEXP(newExp, true, false)
-						partyExp = int32(float64(newExp) * 0.25) // TODO: party exp needs to be properly calculated
+						var partyExp int32
+
+						if dmg == v.maxHP {
+							plr.giveEXP(v.exp, true, false)
+							partyExp = int32(float64(v.exp) * 0.25)
+						} else if float64(dmg)/float64(v.maxHP) > 0.60 {
+							plr.giveEXP(v.exp, true, false)
+							partyExp = int32(float64(v.exp) * 0.25)
+						} else {
+							newExp := int32(float64(v.exp) * 0.25)
+
+							if newExp == 0 {
+								newExp = 1
+							}
+
+							plr.giveEXP(newExp, true, false)
+							partyExp = int32(float64(newExp) * 0.25)
+						}
+
+						if plr.party != nil {
+							plr.party.giveExp(plr.ID, partyExp, true)
+						}
 					}
 
-					if plr.party != nil {
-						// TODO: check level difference is appropriate
-						plr.party.giveExp(plr.ID, partyExp, true)
-					}
+					damager.onMobKilled(v.id)
 				}
 
-				// quest mob logic
-
-				// on die logic
 				for _, id := range v.revives {
 					spawnID, err := pool.nextMobID()
 
@@ -424,59 +420,55 @@ func (pool *lifePool) mobDamaged(poolID int32, damager *Player, dmg ...int32) {
 				}
 
 				pool.removeMob(v.spawnID, 0x1)
-				damager.onMobKilled(v.id)
 
-				if dropEntry, ok := dropTable[v.id]; ok {
-					var mesos int32
-					drops := make([]Item, 0, len(dropEntry))
+				if damager != nil {
+					if dropEntry, ok := dropTable[v.id]; ok {
+						var mesos int32
+						drops := make([]Item, 0, len(dropEntry))
 
-					for _, entry := range dropEntry {
-						if entry.IsMesos {
-							mesos = randRangeInclusive(pool.rNumber, entry.Min, entry.Max)
-							continue
-						}
-
-						// Quest-gated Item: only allow if killer has quest active
-						// This should probably be hidden from instance and only viewable to Player
-						if entry.QuestID != 0 && !damager.allowsQuestDrop(entry.QuestID) {
-							continue
-						}
-
-						if !rollDrop(pool.rNumber, entry.Chance, pool.dropPool.rates.drop) {
-							continue
-						}
-
-						var amount int16 = 1
-						minAmt := entry.Min
-						maxAmt := entry.Max
-						if maxAmt != 1 {
-							val := randRangeInclusive(pool.rNumber, minAmt, maxAmt)
-							if val > math.MaxInt16 {
-								amount = math.MaxInt16
-							} else if val < 1 {
-								amount = 1
-							} else {
-								amount = int16(val)
+						for _, entry := range dropEntry {
+							if entry.IsMesos {
+								mesos = randRangeInclusive(pool.rNumber, entry.Min, entry.Max)
+								continue
 							}
+
+							if entry.QuestID != 0 && !damager.allowsQuestDrop(entry.QuestID) {
+								continue
+							}
+
+							if !rollDrop(pool.rNumber, entry.Chance, pool.dropPool.rates.drop) {
+								continue
+							}
+
+							var amount int16 = 1
+							minAmt := entry.Min
+							maxAmt := entry.Max
+							if maxAmt != 1 {
+								val := randRangeInclusive(pool.rNumber, minAmt, maxAmt)
+								if val > math.MaxInt16 {
+									amount = math.MaxInt16
+								} else if val < 1 {
+									amount = 1
+								} else {
+									amount = int16(val)
+								}
+							}
+
+							newItem, err := CreateItemFromID(entry.ItemID, amount)
+							if err != nil {
+								log.Println("Failed to create drop for mobID:", v.id, "with error:", err)
+								continue
+							}
+							drops = append(drops, newItem)
 						}
 
-						newItem, err := CreateItemFromID(entry.ItemID, amount)
-						if err != nil {
-							log.Println("Failed to create drop for mobID:", v.id, "with error:", err)
-							continue
-						}
-						drops = append(drops, newItem)
+						pool.dropPool.createDrop(dropSpawnNormal, dropFreeForAll, int32(damager.rates.mesos*float32(mesos)), v.pos, true, 0, 0, drops...)
 					}
-
-					// TODO: droppool type determination between DropTimeoutNonOwner and DropTimeoutNonOwnerParty
-					pool.dropPool.createDrop(dropSpawnNormal, dropFreeForAll, int32(damager.rates.mesos*float32(mesos)), v.pos, true, 0, 0, drops...)
-
-					// If has hp bar: remove
 				}
 
 				if v.spawnInterval > 0 {
 					for i, k := range pool.spawnableMobs {
-						if k.id == v.id { // if this needs strengthening then add a spawn pos check
+						if k.id == v.id {
 							pool.spawnableMobs[i].timeToSpawn = time.Now().Add(time.Millisecond * time.Duration(v.spawnInterval))
 							break
 						}
@@ -484,6 +476,72 @@ func (pool *lifePool) mobDamaged(poolID int32, damager *Player, dmg ...int32) {
 				}
 			}
 			break
+		}
+	}
+}
+
+func (pool *lifePool) handleMobSummon(summoner *monster, skillLevel byte, skillData nx.MobSkill) {
+	if len(skillData.MobID) == 0 {
+		return
+	}
+
+	minY := summoner.pos.y - 100
+	maxY := summoner.pos.y + 100
+	minX := summoner.pos.x - 200
+	maxX := summoner.pos.x + 200
+
+	randX := minX
+	if maxX > minX {
+		randX = minX + int16(pool.rNumber.Intn(int(maxX-minX)))
+	}
+
+	for i, mobID := range skillData.MobID {
+		summonY := minY
+		if maxY > minY {
+			summonY = minY + int16(pool.rNumber.Intn(int(maxY-minY)))
+		}
+
+		offset := int16(0)
+		if i > 0 {
+			if i%2 == 1 {
+				offset = int16(35 * (i + 1) / 2)
+			} else {
+				offset = int16(-40 * (i / 2))
+			}
+		}
+		summonX := randX + offset
+		spawnPos := newPos(summonX, summonY, summoner.pos.foothold)
+
+		err := pool.spawnMobFromID(
+			int32(mobID),
+			spawnPos,
+			true,
+			false,
+			false,
+			summoner.spawnID,
+		)
+
+		if err != nil {
+			continue
+		}
+
+		var spawnedMob *monster
+		var highestID int32
+		for id, m := range pool.mobs {
+			if id > highestID {
+				highestID = id
+				spawnedMob = m
+			}
+		}
+
+		if spawnedMob != nil {
+			summonType := int8(skillData.SummonEffect)
+			if summonType == 0 {
+				summonType = -2
+			}
+			spawnedMob.summonType = summonType
+			spawnedMob.summoner = summoner.controller
+			spawnedMob.faceLeft = summoner.faceLeft
 		}
 	}
 }
